backend/pyproject.toml
[tool.poetry]
name = "scraping-wizard-backend"
version = "0.1.0"
description = "Backend API for a no-code web-scraping platform"
authors = ["Scraping Wizard Team"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
supabase = "^2.0.0"
sqlalchemy = "^2.0.21"
asyncpg = "^0.28.0"
celery = "^5.3.4"
redis = "^5.0.1"
aioredis = "^2.0.1"
sse-starlette = "^1.8.2"
playwright = "^1.38.0"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 